Add base model for duplicate functions
