+++ models/GATStack.py
@@ -2,10 +2,12 @@ import torch
 import torch.nn.functional as F
 from torch.nn import ModuleList
 from torch.nn import Sequential, ReLU, Linear
-from torch_geometric.nn import GATConv, BatchNorm, global_mean_pool
+from torch_geometric.nn import GATConv, BatchNorm
+
+from .Base import Base
 
 
-class GATStack(torch.nn.Module):
+class GATStack(Base):
     def __init__(
         self,
         input_dim,
@@ -55,27 +57,6 @@ class GATStack(torch.nn.Module):
             ReLU(),
             Linear(25, output_dim),
         )
-
-    def forward(self, data):
-        x, edge_index, batch = data.x, data.edge_index, data.batch
-        for conv, batch_norm in zip(self.convs, self.batch_norms):
-            x = F.relu(batch_norm(conv(x, edge_index)))
-        x = global_mean_pool(x, batch)
-        return self.mlp(x)
-
-    def loss(self, pred, value):
-        pred_shape = pred.shape
-        value_shape = value.shape
-        if pred_shape != value_shape:
-            value = torch.reshape(value, pred_shape)
-        return F.l1_loss(pred, value)
-
-    def loss_rmse(self, pred, value):
-        pred_shape = pred.shape
-        value_shape = value.shape
-        if pred_shape != value_shape:
-            value = torch.reshape(value, pred_shape)
-        return torch.sqrt(F.mse_loss(pred, value))
 
     def __str__(self):
         return "GATStack"
