+++ models/MFCStack.py
@@ -4,8 +4,10 @@ from torch.nn import ModuleList
 from torch.nn import Sequential, ReLU, Linear
 from torch_geometric.nn import MFConv, BatchNorm, global_mean_pool
 
+from .Base import Base
 
-class MFCStack(torch.nn.Module):
+
+class MFCStack(Base):
     def __init__(
         self,
         input_dim,
@@ -45,27 +47,6 @@ class MFCStack(torch.nn.Module):
             ReLU(),
             Linear(25, output_dim),
         )
-
-    def forward(self, data):
-        x, edge_index, batch = data.x, data.edge_index, data.batch
-        for conv, batch_norm in zip(self.convs, self.batch_norms):
-            x = F.relu(batch_norm(conv(x, edge_index)))
-        x = global_mean_pool(x, batch)
-        return self.mlp(x)
-
-    def loss(self, pred, value):
-        pred_shape = pred.shape
-        value_shape = value.shape
-        if pred_shape != value_shape:
-            value = torch.reshape(value, pred_shape)
-        return F.l1_loss(pred, value)
-
-    def loss_rmse(self, pred, value):
-        pred_shape = pred.shape
-        value_shape = value.shape
-        if pred_shape != value_shape:
-            value = torch.reshape(value, pred_shape)
-        return torch.sqrt(F.mse_loss(pred, value))
 
     def __str__(self):
         return "MFCStack"
