+++ models/__init__.py
@@ -1,3 +1,4 @@
+from .Base import Base
 from .GATStack import GATStack
 from .GINStack import GINStack
 from .PNNStack import PNNStack
